requirements.txt
django
django-extensions
django-flashpolicies
django-piston
docutils
httplib2
python-memcached
oauth
oauth2
pysolr
twisted
